tar vulnerability fix
Vulnerable versions: < 4.4.2
Patched version: 4.4.2
A vulnerability was found in node-tar before version 4.4.2. An Arbitrary
File Overwrite issue exists when extracting a tarball containing a
hardlink to a file that already exists on the system, in conjunction
with a later plain file with the same name as the hardlink. This plain
file content replaces the existing file content.